backend/requirements.txt
# Backend dependencies (Docker). Install via backend image / requirements-dev for tests/lint.

flask>=3.0.0
python-dotenv>=1.0.0

# Database (MySQL, 3NF). Code uses mysql.connector (this package); pymysql is optional.
sqlalchemy>=2.0.0
mysql-connector-python>=8.0.0
pymysql>=1.1.0

# Auth (JWT, password hashing)
pyjwt>=2.8.0
bcrypt>=4.0.0

# PDF / text / DOCX (syllabus parsing)
pypdf>=3.0.0
pdfplumber>=0.10.0

# LLM (hybrid parser - OpenAI)
openai>=1.0.0
python-dateutil>=2.8.0
python-docx>=1.0.0

# CORS
flask-cors>=4.0.0

# Production WSGI (Render, etc.)
gunicorn>=21.0.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
